added minimap button grabber
This minimap button grabber is not perfect but it works. The only problem I faced was if I set opacity to 0 and only to 1 if hovered, after zone change the opacity is always zero no matter what I did.
I will work in this problem later, for now you can either use it as it is or disable it until fully functional.